+++ srcs/ping_send.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <netdb.h>
 #include <sys/time.h>
-#include "utils_time.h"
+#include "utils.h"
 
 static void	_send_echo(char *msg, t_ping *ping)
 {
